Features/gpu fft

+++ SpectrumAnalyzer/Renderer/WaterfallRepresentation.cs
@@ -34,7 +34,7 @@ public class WaterfallRepresentation : FftRepresentation<WaterfallDrawingPropert
         _cyclicScreenBufferMemory = new Memory<byte>(_cyclicScreenBuffer);
     }
 
-    protected override void Draw(Memory<Point> generatePoints, Span<double> magnitudes, Span<double> freqs)
+    protected override void Draw(Memory<Point> generatePoints, Span<float> magnitudes, Span<float> freqs)
     {
         GeneratePoints(generatePoints.Span, magnitudes, freqs);
